Fix csv_summarize crash on CSV where NHS numbers don't have spaces
