+++ project/npda/tests/test_csv_upload.py
@@ -42,16 +42,6 @@ ALDER_HEY_PZ_CODE = "PZ074"
 
 
 @pytest.fixture
-def dummy_sheets_folder(request):
-    return request.config.rootdir / 'project' / 'npda' / 'dummy_sheets'
-
-@pytest.fixture
-def dummy_sheet_csv(dummy_sheets_folder):
-    file = dummy_sheets_folder / 'dummy_sheet.csv'
-    with open(file, 'r') as f:
-        return f.read()
-
-@pytest.fixture
 def valid_df(dummy_sheets_folder):
     return read_csv(dummy_sheets_folder / 'dummy_sheet.csv')
